Use `Int32` for `GitBlobFilterOptions.version` and version macro
The `version` field of `git_blob_filter_options` uses a type of `Int32`,
which also aligns with `GIT_BLOB_FILTER_OPTIONS_VERSION`, but the
`version` parameter of `git_blob_filter_options()` uses `UInt32`. To
minimize integer casting, use `Int32` for `GitBlobFilterOptions.version`
and for the `GitBlobFilterOptionsVersion` macro.

`gitBlobFilterOptionsInit(opts:version:)` continues using `UInt32` to
match the C function's parameter type.
`GitBlobFilterOptions.withCValue(_:)` validates that its `version`
property is not negative before casting to `UInt32`, and will throw an
error otherwise.

+++ Tests/swift-libgit2Tests/Blob-Tests.swift
@@ -355,7 +355,7 @@ final class BlobTests: XCTestCaseStopOnFail
         {
             cBlobFilterOptions in
             
-            XCTAssertEqual(cBlobFilterOptions.pointee.version, Int32(gitBlobFilterOptionsVersion))
+            XCTAssertEqual(cBlobFilterOptions.pointee.version, gitBlobFilterOptionsVersion)
             XCTAssertEqual(GitBlobFilterFlagT(rawValue: cBlobFilterOptions.pointee.flags), .gitBlobFilterCheckForBinary)
             XCTAssertNil(cBlobFilterOptions.pointee.commit_id)
             XCTAssertZeroOID(GitOID(cValue: cBlobFilterOptions.pointee.attr_commit_id))
@@ -371,7 +371,7 @@ final class BlobTests: XCTestCaseStopOnFail
         let blobFilterOptionsInitResult: GitErrorCode
             = gitBlobFilterOptionsInit(
                 opts:       &blobFilterOptions,
-                version:    gitBlobFilterOptionsVersion
+                version:    UInt32(gitBlobFilterOptionsVersion)
             )
         
         XCTAssertOK(blobFilterOptionsInitResult)
@@ -381,7 +381,7 @@ final class BlobTests: XCTestCaseStopOnFail
     
     func testGitBlobFilterOptionsVersion() throws
     {
-        XCTAssertEqual(Int32(gitBlobFilterOptionsVersion), GIT_BLOB_FILTER_OPTIONS_VERSION)
+        XCTAssertEqual(gitBlobFilterOptionsVersion, GIT_BLOB_FILTER_OPTIONS_VERSION)
     }
